permission_granted? now either returns a boolean or raises
Returning a boolean is more idiomatic for functions that end in a `?`, and it
also makes permission checks in the code easier to read. However, we can't then
rely on returning an error tuple when one of the ids being checked is invalid,
because the tuple would be interpreted as a truthy value and result in a
false-positive if, for example, there is a typo in the name of the permission
being checked. I decided that raising an exception is the right way to go,
becuase the library should expect that the dependant application has already
verified in its own manner that the ids being checked are valid.